+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [1.1.0] - 2023-08-01
+
+### Added
+
+- New ```ScenesManagerConfigConstants``` to keep the scenes keys as enums;
+- Editor for ```ScenesManagerConfigConstants```;
+
+### Changed
+
+- Left only ```LoadScene(Enum sceneKey)``` and ```AsyncOperation LoadSceneAsync(Enum sceneKey)```;
+
+### Removed
+
+- Flag ```async``` from the ```ScenesManagerConfigData```;
+
 ## [0.0.2] - 2021-04-24
 - Test verdaccio Slack notification.